Break out deref.hpp
